Create README.md
a directory website for local businesses in the city you live in. The website have the functionality to sort the businesses based on customer reviews.  as many classes of businesses a possible, for example: restaurants, auto services, home services to name a few. the functionality to categories these businesses based on the neighborhood they are located in. an effect like animations and transitions to the page to make it looks visually appealing.